+++ wayland-backend/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Additions
+- backend: Added a `destroy_object` method
+
 ## 0.3.10 -- 2025-04-29
 
 #### Bugfixes