+++ wayland-backend/CHANGELOG.md
@@ -5,9 +5,11 @@
 #### Bugfixes
 
 - backend/server/rs: Send protocol error if method requires newer protocol version
+
 ### Additions
 
 - backend: Added a `destroy_object` method
+- backend/server: Remove registry from `known_registries` if destructor is called
 
 ## 0.3.10 -- 2025-04-29